+++ src/library/stats/R/ar.yw.mts.R
@@ -51,7 +51,7 @@ function (x, aic = TRUE, order.max = NULL, na.action = na.fail,
             pacf = double((1 + order.max) * nser * nser),
             var = double((1 + order.max) * nser * nser),
             aic = double(1 + order.max),
-            order = integer(1),
+            order = integer(1L),
             as.integer(aic))
     partialacf <- aperm(array(z$pacf, dim = c(nser, nser, order.max +
         1)), c(3, 2, 1))[-1, , , drop = FALSE]
@@ -64,10 +64,10 @@ function (x, aic = TRUE, order.max = NULL, na.action = na.fail,
     if (order > 0) {
         ar <- -aperm(array(z$coefs, dim = c(nser, nser, order.max +
             1)), c(3, 2, 1))[2:(order + 1), , , drop = FALSE]
-        for (i in 1:order) resid[-(1:order), ] <- resid[-(1:order),
+        for (i in 1L:order) resid[-(1L:order), ] <- resid[-(1L:order),
             ] - x[(order - i + 1):(n.used - i), ] %*% t(ar[i,
             , ])
-        resid[1:order, ] <- NA
+        resid[1L:order, ] <- NA
     }
     else ar <- array(dim = c(0, nser, nser))
     var.pred <- var.pred[order + 1, , , drop = TRUE] * n.used/(n.used -
@@ -80,7 +80,7 @@ function (x, aic = TRUE, order.max = NULL, na.action = na.fail,
     colnames(resid) <- snames
     dimnames(ar) <- list(seq_len(order), snames, snames)
     dimnames(var.pred) <- list(snames, snames)
-    dimnames(partialacf) <- list(1:order.max, snames, snames)
+    dimnames(partialacf) <- list(1L:order.max, snames, snames)
     res <- list(order = order, ar = ar, var.pred = var.pred,
         x.mean = x.mean, aic = xaic, n.used = n.used, order.max = order.max,
         partialacf = partialacf, resid = resid, method = "Yule-Walker",
